rotem_sela_top/backend-docker/rotem_sela_backend/test.py
import v4l2
import fcntl
import mmap
import os

def open_device(device_path):
    # Open the video device
    fd = os.open(device_path, os.O_RDWR, 0)
    return fd

def init_device(fd, width, height):
    # Set the video format (Assuming MJPEG here)
    format = v4l2.v4l2_format()
    format.type = v4l2.V4L2_BUF_TYPE_VIDEO_CAPTURE
    format.fmt.pix.width = width
    format.fmt.pix.height = height
    format.fmt.pix.pixelformat = v4l2.V4L2_PIX_FMT_MJPEG
    format.fmt.pix.field = v4l2.V4L2_FIELD_INTERLACED
    fcntl.ioctl(fd, v4l2.VIDIOC_S_FMT, format)

def request_buffers(fd, num_buffers):
    # Request buffers from the device
    req = v4l2.v4l2_requestbuffers()
    req.type = v4l2.V4L2_BUF_TYPE_VIDEO_CAPTURE
    req.memory = v4l2.V4L2_MEMORY_MMAP
    req.count = num_buffers
    fcntl.ioctl(fd, v4l2.VIDIOC_REQBUFS, req)

def mmap_buffers(fd, num_buffers):
    # Map the buffers
    buffers = []
    for index in range(num_buffers):
        buf = v4l2.v4l2_buffer()
        buf.type = v4l2.V4L2_BUF_TYPE_VIDEO_CAPTURE
        buf.memory = v4l2.V4L2_MEMORY_MMAP
        buf.index = index
        fcntl.ioctl(fd, v4l2.VIDIOC_QUERYBUF, buf)

        # Map the buffer to user space
        buffer_data = mmap.mmap(
            fd, buf.length, mmap.MAP_SHARED, mmap.PROT_READ | mmap.PROT_WRITE, offset=buf.m.offset
        )
        buffers.append((buf, buffer_data))

    return buffers

def start_capture(fd):
    # Start streaming
    buf_type = v4l2.v4l2_buf_type(v4l2.V4L2_BUF_TYPE_VIDEO_CAPTURE)
    fcntl.ioctl(fd, v4l2.VIDIOC_STREAMON, buf_type)


def capture_frame(fd, buffers):
    # Queue the buffer for capture
    buf = v4l2.v4l2_buffer()
    buf.type = v4l2.V4L2_BUF_TYPE_VIDEO_CAPTURE
    buf.memory = v4l2.V4L2_MEMORY_MMAP
    fcntl.ioctl(fd, v4l2.VIDIOC_QBUF, buf)

    # Dequeue the buffer after it has been filled
    fcntl.ioctl(fd, v4l2.VIDIOC_DQBUF, buf)

    # Get the frame data
    frame_data = buffers[buf.index][1]

    return frame_data

def stop_capture(fd):
    # Stop streaming
    buf_type = v4l2.v4l2_buf_type(v4l2.V4L2_BUF_TYPE_VIDEO_CAPTURE)
    fcntl.ioctl(fd, v4l2.VIDIOC_STREAMOFF, buf_type)


def close_device(fd):
    # Close the video device
    os.close(fd)

if __name__ == "__main__":
    # Replace '/dev/video0' with the path to your video device
    device_path = '/dev/video2'

    # Set the desired width and height
    width, height = 640, 480

    # Number of buffers to use
    num_buffers = 4

    # Open the video device
    fd = open_device(device_path)

    # Initialize the video device
    init_device(fd, width, height)

    # Request buffers from the device
    request_buffers(fd, num_buffers)

    # Map the buffers
    buffers = mmap_buffers(fd, num_buffers)

    # Start streaming
    start_capture(fd)

    try:
        # Capture a frame
        frame_data = capture_frame(fd, buffers)

        # Process the frame_data as needed (e.g., save to a file)
        # ...

    finally:
        # Stop streaming
        stop_capture(fd)

        # Close the video device
        close_device(fd)

