rotem_sela_top/backend-docker/rotem_sela_backend/requirements.txt
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
chardet==3.0.4
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==3.1.0
Flask-WTF==0.12
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
h11==0.14.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
opencv-python==4.8.1.78
python-engineio==4.8.0
python-socketio==5.10.0
pyusb==1.2.1
requests==2.23.0
simple-websocket==1.0.0
six==1.16.0
tornado==6.0.4
urllib3==1.25.11
v4l2py==2.3.0
websockets==7.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.0
zope.event==5.0
zope.interface==6.1
